fix multiple indexes, type:module
