+++ package.json
@@ -4,12 +4,13 @@
     "node-static": "^0.7.11"
   },
   "name": "Incognito",
+  "type" :"module",
   "version": "1.0.0",
   "description": "Incognito with bare server",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node ."
+    "start": "node index.js"
   },
   "keywords": [
     "titaniumnetwork",
